refactor: update webpack to use browsersync for more stable server reloads; break out webpack config handling verbosity into separate include
